move tileJSON support from tile-layer to mvt-layer
